[next.js learn] Streaming
https://nextjs.org/learn/dashboard-app/streaming
