+++ nextjs-dashboard/README.md
@@ -200,3 +200,20 @@ __Using SQL__
 - **사용자별 콘텐츠** - 대시보드나 사용자 프로필과 같은 개인화된 콘텐츠를 제공하고 사용자 상호 작용에 따라 데이터를 업데이트하는 것이 더 쉬워집니다.
 - **요청 시간 정보** - 동적 렌더링을 사용하면 쿠키 또는 URL 검색 매개변수와 같이 요청 시점에만 알 수 있는 정보에 액세스할 수 있습니다.
 
+---
+
+## [Streaming](https://nextjs.org/learn/dashboard-app/streaming)
+
+이전 장에서는 Next.js의 다양한 렌더링 방법에 대해 배웠습니다. 또한 느린 데이터 가져오기가 애플리케이션의 성능에 어떤 영향을 미칠 수 있는지에 대해서도 설명했습니다. 느린 데이터 요청이 있을 때 사용자 환경을 개선할 수 있는 방법을 살펴보겠습니다.
+
+### What is streaming?
+
+스트리밍은 경로를 더 작은 '청크'로 나누고 준비되는 대로 서버에서 클라이언트로 점진적으로 스트리밍할 수 있는 데이터 전송 기술입니다.
+
+스트리밍을 사용하면 느린 데이터 요청으로 인해 전체 페이지가 차단되는 것을 방지할 수 있습니다. 이를 통해 사용자는 모든 데이터가 로드될 때까지 기다리지 않고도 페이지의 일부를 보고 상호 작용할 수 있으며, UI가 사용자에게 표시될 수 있습니다.
+
+스트리밍은 각 컴포넌트를 청크로 간주할 수 있기 때문에 React의 컴포넌트 모델에서 잘 작동합니다.
+
+Next.js에서 스트리밍을 구현하는 방법에는 두 가지가 있습니다:
+1. 페이지 수준에서 `loading.tsx` 파일(`<Suspense>`가 생성됨)을 사용합니다.
+2. 컴포넌트 수준에서 `<Suspense>`를 사용하면 더욱 세밀하게 제어할 수 있습니다.
